requirements.txt
#
####### example-requirements.txt #######
#
###### Requirements without Version Specifiers ######
pillow
arcade
#